requirements.txt
gdown==4.4.0
marshmallow==3.14.1
marshmallow_dataclass==8.5.8
pandas==1.1.5
pytest==7.0.1
PyYAML==6.0
scikit_learn==1.0.2
